requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
apscheduler==3.10.4
email-validator==2.1.0.post1
aiosmtplib==2.0.2
pydantic==2.5.1
pydantic[email]
pytest>=7.4.3
pytest-asyncio>=0.23.2
pytest-cov>=5.0.0
pytest-html>=4.1.1
requests>=2.26.0
playwright>=1.39.0
selenium>=4.15.2
appium-python-client>=2.11.1
fastapi-mail==1.4.1 